Changed pet.py to use PyQt5 Changed requirements.txt Added bin\angkal for windows
