+++ bin/angkal.bat
@@ -1,0 +1,2 @@
+@echo off
+python pet.py
